Move pause/start/stop to toolbars, remove ContextAwareAction which doesn't seem to do anything.
